Allow transcoder to run concurently
