--- transcoder/transcoder/stream.go
@@ -1,9 +1,0 @@
-package transcoder
-
-type TranscodeStream struct {
-	File    FileStream
-	Clients []string
-	// true if the segment at given index is completed/transcoded, false otherwise
-	segments []bool
-	// TODO: add ffmpeg process
-}
